requirements.txt
langchain>=0.1.0
langgraph>=0.0.15
langchain-openai>=0.0.2
langchain-community>=0.0.10
langchain-chroma>=0.0.1
chromadb>=0.4.22
python-dotenv>=1.0.0
pydantic>=2.5.0
streamlit>=1.30.0
pandas>=2.0.0
openpyxl>=3.0.0
tiktoken>=0.5.1
requests>=2.31.0
sentence-transformers>=2.2.2 